Fix charging seed: round() needs a numeric arg, not double precision
energy_kwh's ::numeric cast only wrapped the loss-factor term, not the
whole expression — multiplying that against the (numeric) battery-delta
terms still resolves to double precision overall in Postgres's mixed
numeric/float8 arithmetic, same class of bug as the earlier scorecard
achieved_points line. round(double precision, int) has no 2-arg
overload. Cast the whole expression instead, matching every other
round() call in this file.

+++ supabase/seed/test_data.sql
@@ -732,7 +732,7 @@ shaped as (
 ),
 energy_calc as (
   select *,
-    round(300 * (battery_end - battery_start) / 100.0 * (1.05 + random() * 0.07), 2) as energy_kwh
+    round((300 * (battery_end - battery_start) / 100.0 * (1.05 + random() * 0.07))::numeric, 2) as energy_kwh
   from shaped
 ),
 final as (
